requirements.txt
aiogram==3.13.0
aladhan.py
python-dotenv
apscheduler
